+++ coolprompt/utils/var_validation.py
